move icons dir
