+++ src/icons/ExplosionIcon.tsx
