+++ src/icons/NewTabIcon.tsx
