+++ src/icons/RemoveFromListIcon.tsx
