+++ src/icons/SpinnerIcon.tsx
